+++ tests/testthat/test-ChoicesFilterState.R
@@ -89,3 +89,38 @@ testthat::test_that("set_keep_na_reactive throws error if input is not logical",
   testthat::expect_error(filter_state$set_keep_na_reactive("TRUE"))
   testthat::expect_error(filter_state$set_keep_na_reactive(1))
 })
+
+testthat::test_that(
+  "ChoicesFilterState$is_any_filtered works properly when NA is present in data",
+  code = {
+    filter_state <- teal:::ChoicesFilterState$new(
+      c(LETTERS[1:2], NA),
+      varname = "x",
+      input_dataname = as.name("data"),
+      extract_type = character(0)
+    )
+    isolate(filter_state$set_keep_na(FALSE))
+    isolate(filter_state$set_selected(LETTERS[1:2]))
+    testthat::expect_true(
+      isolate(filter_state$is_any_filtered())
+    )
+
+    isolate(filter_state$set_keep_na(TRUE))
+    isolate(filter_state$set_selected(LETTERS[1:2]))
+    testthat::expect_false(
+      isolate(filter_state$is_any_filtered())
+    )
+
+    isolate(filter_state$set_selected(LETTERS[1]))
+    isolate(filter_state$set_keep_na(TRUE))
+    testthat::expect_true(
+      isolate(filter_state$is_any_filtered())
+    )
+
+    isolate(filter_state$set_selected(LETTERS[1]))
+    isolate(filter_state$set_keep_na(FALSE))
+    testthat::expect_true(
+      isolate(filter_state$is_any_filtered())
+    )
+  }
+)